refactor(rentals): split return finalization flow
